Move and simplify superposition JSON test
This test was previously working on my machines but failing on the
Travis-CI build because it was producing slightly different
floating-point numbers in the JSON output.

This change simplifies the test to use whole numbers in the translation
and rotation matrix.

It also moves the test to a separate superposition_io_test.cpp
translation unit.