Fix bug: phi/psi angle's acos of <-1 (float rndng)
Resolves #7

Phi/psi angles broke on residues 25 and 26 of chain A of 1cn1 because
the values conspired with floating-point rounding errors to give a value
of less than -1 with which to call acos. Now fixed by clamping.

Also added a testcase that replicated the failure (before I fixed it).

Thanks to Natalie for finding this bug and to Ian for writing it up as a
GitHub issue.