Bugfix: fix sec overrun problem flagged by Natalie
Natalie flagged up a bug she saw when using cath-ssap on 4ba8A00 versus
1h3tB01 that caused a segmentation fault.

This was caused by the problem we've previously seen that the sec files'
residue numbering doesn't reliably tally with that in the DSSP/wolf
files. This can cause one of the secondary structures to overrun the end
of the residue records. The code was already doing some sensible
handling of any overruns of the end of the secondary structure. However
in 1h3tB01, the secondary structure's *start* also overruns and the code
wasn't handling this and so seg-faulted.

The code has now been altered to cap the start at the end of the list of
residues.

I have sanity-checked this change with a run of the regression_test in
cath-tools-supplementary