Bugfix: abide PDB conflicts for same residue
Natalie flagged up a bug she saw when using cath-ssap on 2tnfB00 versus
4tswB00 that broke the comparison.

This turned out to be because 4tsw (which is obsolete and superseded by
5tsw) has conflicting ATOM records for the same residue 103 on chain B
(and on other chains). Though this shouldn't happen in PDBs, it turns
out that it does so this change demotes this situation from an error to
a warning. The code now ignore the latter residue's records when parsing
PDBs or DSSPs.