Bugfix: workaround DSSP not knowing PYL/O, SEC/U
Natalie flagged up a bug she saw when using cath-ssap on 1nsqC00 versus
1paeX00 that broke the comparison.

The error was that the code detected a conflict between the amino acid
for a specific residue in the PDB / DSSP files. This looks to just be a
case of DSSP not recognising PYL/O and SEC/U and hence converting to
UNK/X. We've decided to specifically test for this type of conflict and
just warn rather than error-out in this case.