Fix bug: handle residue 0 after null in DSSP files
There was previously a subtle bug here which meant the code mistakenly
skipped any residue 0 that immediately followed a null residue record in
a DSSP file. This absence caused the code to fall over later on when
tallying PDB residues with DSSP residues.

This affected PDB chains 1k7kA and 1zkpB.