Change full_hit to store hit_extras info
This provides a consistent way for extra pieces of information to be
added to a full_hit. In particular, it allows HMMER conditional evalues
and independent evalues to be stored.

The alnd_rgn_vec has been moved inside this hit_extras store and is
currently just stored as a string (though this could be changed if
needed).