Try to improve Ubuntu Travis-CI exes' back-compat
@shouldsee has spotted (#42) that the executables from the current
Travis-CI Ubuntu build don't run on an out-of-the box Ubuntu 16.04
because it complains about not having GLIBCXX_3.4.22.

I think this is caused by the `.travis.yml` getting a more recent GCC
than the standard from `ppa:ubuntu-toolchain-r/test` and that GCC
bringing in a more recent `libstdc++`.

I don't want to spend lots of time trying to get a perfect solution but
it'd be good to at least reduce the minimum Ubuntu requirement
(particularly to 16.04 LTS).

This commit changes the compiler:
 * from g++ 6.3.0, which creates an exe that requires GLIBCXX_3.4.22,
 * added in Ubuntu 16.10
 * to g++ 4.9.4, which (hopefully) creates an exe that requires
 * GLIBCXX_3.4.20, added in 14.10 (I think; definitely by 15.04)

For future reference: the intermediate is GCC 5.1.0 / GLIBCXX_3.4.21,
added in 15.10