Improve handling of external info from build
Improve handling of git version and also make CMAKE_SOURCE_DIR and
CMAKE_BINARY_DIR available in the code so it can be used so tests know
where to find the test data files in the source directory.

As part of that, make source/CMakeLists.txt correctly handle modules
that have no tests.